dsp16xx.c (print_operand_address): Fix format specifier warning.
	* dsp16xx.c (print_operand_address): Fix format specifier warning.
	* dsp16xx.h (INCLUDE_DEFAULTS): Add missing initializers.

From-SVN: r66882

+++ gcc/config/dsp16xx/dsp16xx.c
@@ -1938,7 +1938,7 @@ print_operand_address(file, addr)
       
     default:
       if (FITS_5_BITS (addr))
-	fprintf (file, "*(0x%x)", (INTVAL (addr) & 0x20));
+	fprintf (file, "*(0x%x)", (int)(INTVAL (addr) & 0x20));
       else
 	output_addr_const (file, addr);
     }
